:truck: Split entry builder into a module directory
